[Refactor] (Harness): 将定时器功能整合到Harness类中并移除冗余代码
将原本独立的HarnessTimer类功能迁移到Harness类内部
移除冗余的C语言头文件包含和全局LED定义
简化消息处理模块对定时器的调用方式
